chore: fix lints

+++ crates/store/src/server/api.rs
@@ -244,7 +244,7 @@ impl api_server::Api for StoreApi {
             .notes
             .into_iter()
             .map(|note| NoteSyncRecord {
-                note_index: note.note_index.to_absolute_index() as u32,
+                note_index: note.note_index.to_absolute_index(),
                 note_id: Some(note.note_id.into()),
                 metadata: Some(note.metadata.into()),
                 merkle_path: Some(note.merkle_path.into()),
